create story component with external button to minimize
